requirements.txt
absl-py==2.2.2
astunparse==1.6.3
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.4.0
google-auth==2.40.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio @ file:///C:/b/abs_4cup9zfkvm/croot/imageio_1738160462075/work
imageio-ffmpeg==0.6.0
joblib==1.5.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.1
namex==0.0.9
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
protobuf==3.19.6
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.10.1
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==1.12.0
torchvision==0.13.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wrapt==1.17.2
